+++ examples/lm3s6965/examples/async-timeout.rs
@@ -8,13 +8,13 @@
 
 use cortex_m_semihosting::{debug, hprintln};
 use panic_semihosting as _;
-use rtic_monotonics::systick::*;
+use rtic_monotonics::systick::prelude::*;
+systick_monotonic!(Mono, 100);
 
 #[rtic::app(device = lm3s6965, dispatchers = [SSI0, UART0], peripherals = true)]
 mod app {
     use super::*;
     use futures::{future::FutureExt, select_biased};
-    use rtic_monotonics::Monotonic;
 
     #[shared]
     struct Shared {}
@@ -27,8 +27,7 @@ mod app {
     fn init(cx: init::Context) -> (Shared, Local) {
         hprintln!("init");
 
-        let systick_token = rtic_monotonics::create_systick_token!();
-        Systick::start(cx.core.SYST, 12_000_000, systick_token);
+        Mono::start(cx.core.SYST, 12_000_000);
         // ANCHOR_END: init
 
         foo::spawn().ok();
@@ -42,19 +41,19 @@ mod app {
         // Call hal with short relative timeout using `select_biased`
         select_biased! {
             v = hal_get(1).fuse() => hprintln!("hal returned {}", v),
-            _ = Systick::delay(200.millis()).fuse() =>  hprintln!("timeout", ), // this will finish first
+            _ = Mono::delay(200.millis()).fuse() =>  hprintln!("timeout", ), // this will finish first
         }
 
         // Call hal with long relative timeout using `select_biased`
         select_biased! {
             v = hal_get(1).fuse() => hprintln!("hal returned {}", v), // hal finish first
-            _ = Systick::delay(1000.millis()).fuse() =>  hprintln!("timeout", ),
+            _ = Mono::delay(1000.millis()).fuse() =>  hprintln!("timeout", ),
         }
         // ANCHOR_END: select_biased
 
         // ANCHOR: timeout_after_basic
         // Call hal with long relative timeout using monotonic `timeout_after`
-        match Systick::timeout_after(1000.millis(), hal_get(1)).await {
+        match Mono::timeout_after(1000.millis(), hal_get(1)).await {
             Ok(v) => hprintln!("hal returned {}", v),
             _ => hprintln!("timeout"),
         }
@@ -62,20 +61,20 @@ mod app {
 
         // ANCHOR: timeout_at_basic
         // get the current time instance
-        let mut instant = Systick::now();
+        let mut instant = Mono::now();
 
         // do this 3 times
         for n in 0..3 {
             // absolute point in time without drift
             instant += 1000.millis();
-            Systick::delay_until(instant).await;
+            Mono::delay_until(instant).await;
 
             // absolute point in time for timeout
             let timeout = instant + 500.millis();
-            hprintln!("now is {:?}, timeout at {:?}", Systick::now(), timeout);
+            hprintln!("now is {:?}, timeout at {:?}", Mono::now(), timeout);
 
-            match Systick::timeout_at(timeout, hal_get(n)).await {
-                Ok(v) => hprintln!("hal returned {} at time {:?}", v, Systick::now()),
+            match Mono::timeout_at(timeout, hal_get(n)).await {
+                Ok(v) => hprintln!("hal returned {} at time {:?}", v, Mono::now()),
                 _ => hprintln!("timeout"),
             }
         }
@@ -90,7 +89,7 @@ async fn hal_get(n: u32) -> u32 {
     // emulate some delay time dependent on n
     let d = 350.millis() + n * 100.millis();
     hprintln!("the hal takes a duration of {:?}", d);
-    Systick::delay(d).await;
+    Mono::delay(d).await;
     // emulate some return value
     5
 }