+++ rtic/src/lib.rs
@@ -34,7 +34,6 @@
 
 pub use rtic_core::{prelude as mutex_prelude, Exclusive, Mutex};
 pub use rtic_macros::app;
-// pub use rtic_monotonic::{self, Monotonic};
 
 /// module `mutex::prelude` provides `Mutex` and multi-lock variants. Recommended over `mutex_prelude`
 pub mod mutex {