+++ rtic/CHANGELOG.md
@@ -7,11 +7,21 @@ For each category, *Added*, *Changed*, *Fixed* add new entries at the top!
 
 ## [Unreleased]
 
+### Changed
+
+- Remove unused dependency `rtic-monotonics`
+
 ## [v2.1.1] - 2024-03-13
 
 ### Fixed
 
 - **Soundness fix:** `thumbv6` was subject to race in source mask.
+
+## [v2.1.0] - 2024-02-27
+
+### Changed
+
+- Remove unused dependency `rtic-monotonics`
 
 ## [v2.1.0] - 2024-02-27